+++ CHANGELOG.md
@@ -8,6 +8,11 @@ For Keycloak upstream changelog please see https://www.keycloak.org/docs/latest/
 [Keycloak announcement for version 16.1.0](https://www.keycloak.org/2021/12/keycloak-1610-released)
 Full Keycloak upstream jira issue can be shown if filtered by Fix version. 
 
+## [Unreleased]
+
+### Fixed
+- Instant.now().toEpochMilli instead of System.currentTimeMillis [RCIAM-1002](https://jira.argo.grnet.gr/browse/RCIAM-1002)
+
 ## [16.1.0-2.2] - 2022-05-19
 ### Changed
 - Fixed the key use in the SAML Federation metadata (SSPSSODescriptor) xml.